server/requirements.txt
# Miku Call Server Dependencies

# Web framework & WebSocket
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==14.1
python-multipart==0.0.20

# AI & Speech
openai-whisper==20231117
anthropic==0.46.0

# Audio processing
pydub==0.25.1
numpy==2.2.3

# Firebase for push notifications
firebase-admin==6.7.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
